+++ src/main/java/application/rest/WishController.java
@@ -1,7 +1,11 @@
 package application.rest;
 
 import application.persistence.entity.Wish;
+import application.persistence.type.UserRoleEnum;
 import application.rest.domain.WishDTO;
+import application.service.response.ServiceResponse;
+import application.service.response.ServiceResponseStatus;
+import application.service.user.UserService;
 import application.service.wish.WishService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
@@ -16,8 +20,18 @@ public class WishController extends AbstractDatabaseController<Wish, Long, WishD
     @Autowired
     private WishService wishService;
 
+    @Autowired
+    private UserService userService;
+
     @RequestMapping(value = "/customers/{customerId}/wishlist", method = RequestMethod.GET)
     public ResponseEntity<?> readUsersWishes(@PathVariable UUID customerId, Pageable pageable) {
+        ServiceResponse<Boolean> hasRolesResponse = userService.hasCurrentUserDemandedRoles(
+                UserRoleEnum.ROLE_ADMIN
+        );
+        ServiceResponse<Boolean> isCurrentUserResponse = userService.isCurrrentUser(customerId);
+        if (!hasRolesResponse.isSuccessful() && !isCurrentUserResponse.isSuccessful()) {
+            return new ErrorResponseEntity(ServiceResponseStatus.FORBIDDEN);
+        }
         return getPageResponseEntity(
                 wishService.readCustomersWishes(customerId, pageable)
         );
@@ -31,12 +45,20 @@ public class WishController extends AbstractDatabaseController<Wish, Long, WishD
     }
 
     @RequestMapping(value = "/customers/{customerId}/wishlist/{wishId}", method = RequestMethod.GET)
-    public ResponseEntity<?> readAddress(@PathVariable UUID customerId, @PathVariable Long wishId) {
+    public ResponseEntity<?> readWish(@PathVariable UUID customerId, @PathVariable Long wishId) {
+        ServiceResponse<Boolean> isCurrentUserResponse = userService.isCurrrentUser(customerId);
+        if (!isCurrentUserResponse.isSuccessful()) {
+            return new ErrorResponseEntity(isCurrentUserResponse.getStatus());
+        }
         return read(wishId);
     }
 
     @RequestMapping(value = "/customers/{customerId}/wishlist", method = RequestMethod.POST)
     public ResponseEntity<?> createWish(@PathVariable UUID customerId, @RequestBody WishDTO wishDTO) {
+        ServiceResponse<Boolean> isCurrentUserResponse = userService.isCurrrentUser(customerId);
+        if (!isCurrentUserResponse.isSuccessful()) {
+            return new ErrorResponseEntity(isCurrentUserResponse.getStatus());
+        }
         wishDTO.setCustomerUid(customerId);
         return create(wishDTO);
     }
@@ -50,6 +72,10 @@ public class WishController extends AbstractDatabaseController<Wish, Long, WishD
 
     @RequestMapping(value = "/customers/{customerId}/wishlist/{productId}", method = RequestMethod.DELETE)
     public ResponseEntity<?> deleteWish(@PathVariable UUID customerId, @PathVariable UUID productId) {
+        ServiceResponse<Boolean> isCurrentUserResponse = userService.isCurrrentUser(customerId);
+        if (!isCurrentUserResponse.isSuccessful()) {
+            return new ErrorResponseEntity(isCurrentUserResponse.getStatus());
+        }
         return getSimpleResponseEntity(
                 wishService.removeProductFromWishes(customerId, productId)
         );
@@ -66,5 +92,6 @@ public class WishController extends AbstractDatabaseController<Wish, Long, WishD
     public WishService getBaseService() {
         return wishService;
     }
+
 
 }
